Update
I added the missing #include directive for the <vector> header file at the beginning of the code.

I added a Scope object named root and initialized it to the root scope using the Scope::NewRootScope() function.

I passed the root scope object to the ClientSession constructor when creating a new ClientSession object. This is necessary because the ClientSession object needs a scope to be able to access the TensorFlow ops and variables that will be used in the model.

I fixed the declaration of the inputs vector by providing the correct data types for the Tensor objects. The input tensor has shape (4, 10) and data type DT_FLOAT, and the label tensor has shape (4, 1) and data type DT_FLOAT.

I added code to initialize the inputTensor and labelTensor objects with the data from the inputData and outputData vectors, respectively.

I defined the TensorFlow ops for the model, including the placeholders for the input and label tensors, the variables for the weights and biases, and the layers and activations of the model.

I defined the loss function as the mean squared error between the label and the output of the model.

I defined an Adam optimizer to minimize the loss function, and provided it with a list of the trainable variables in the model.

I added a loop to train the model for 100 epochs by running the optimizer and computing the loss at each epoch.

I added code to evaluate the model by running it on the input data and printing the output tensor.